Dot sanity checks.

+++ clients/include/testing_scal.hpp
@@ -31,8 +31,8 @@ hipblasStatus_t testing_scal(const Arguments& argus)
         return HIPBLAS_STATUS_SUCCESS;
     }
 
-    int sizeX = N * incx;
-    U   alpha = argus.get_alpha<U>();
+    size_t sizeX = size_t(N) * incx;
+    U      alpha = argus.get_alpha<U>();
 
     // Naming: dX is in GPU (device) memory. hK is in CPU (host) memory, plz follow this practice
     host_vector<T>   hx(sizeX);
